Move some internal methods to ImageInternal.swift
- `_convoluted`
- `_interpolate`
- `_transformed`
